Update to use urlhelper package
