+++ SFA.DAS.Employer.Shared.UI/IntegrationTests/Tests/FooterTests.cs
@@ -1,4 +1,3 @@
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Xunit;
