+++ SFA.DAS.Employer.Shared.UI/IntegrationTests/linkGeneratorSettings.json
@@ -6,6 +6,7 @@
     "PortalBaseUrl": "https://localhost",
     "ProjectionsBaseUrl": "http://localhost:50770",
     "RecruitBaseUrl": "https://recruit.at-eas.apprenticeships.education.gov.uk",
-    "UsersBaseUrl": "https://at-login.apprenticeships.education.gov.uk"
+    "UsersBaseUrl": "https://at-login.apprenticeships.education.gov.uk",
+    "FinanceBaseUrl": "https://finance.at-eas.apprenticeships.education.gov.uk"
   }
 }
