+++ SFA.DAS.Employer.Shared.UI/SFA.DAS.Employer.Shared.UI/SFA.DAS.Employer.Shared.UI.csproj
@@ -17,6 +17,6 @@
     <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="2.2.0"/>
     <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="2.2.0"/>
     <PackageReference Include="Microsoft.Extensions.Options" Version="2.2.0"/>
-    <PackageReference Include="SFA.DAS.EmployerUrlHelper" Version="2.0.2"/>
+    <PackageReference Include="SFA.DAS.EmployerUrlHelper" Version="3.0.12"/>
   </ItemGroup>
 </Project>
