+++ SFA.DAS.Employer.Shared.UI/SFA.DAS.Employer.Shared.UI/ServiceCollectionExtensions.cs
@@ -3,6 +3,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using SFA.DAS.Employer.Shared.UI.Configuration;
 using SFA.DAS.EmployerUrlHelper;
+using SFA.DAS.EmployerUrlHelper.DependencyResolution;
 
 namespace SFA.DAS.Employer.Shared.UI
 {
@@ -17,7 +18,6 @@ namespace SFA.DAS.Employer.Shared.UI
 
             var linkGenerator = services.BuildServiceProvider().GetService<ILinkGenerator>();
 
-            
             // TODO: Validate configuration values?
             services.Configure<MaPageConfiguration>(configuration.GetSection("SFA.DAS.Employer.Shared.UI:MaPageConfiguration"));
             services.PostConfigure<MaPageConfiguration>(options =>
