+++ SFA.DAS.Employer.Shared.UI/SFA.DAS.Employer.Shared.UI/Views/Shared/_Menu.cshtml
@@ -115,7 +115,7 @@
                         <a class="@BuildNavigationLinkClass(NavigationSection.AccountsHome, selectedSection) mu-home" href=href="@linkGenerator.AccountsLink("AccountsHome", accountId)">Home</a>
                     </li>
                     <li class="das-navigation__list-item">
-                        <a class="@BuildNavigationLinkClass(NavigationSection.AccountsFinance, selectedSection) mu-finance" href="@linkGenerator.AccountsLink("AccountsFinance", accountId)">Finance</a>
+                        <a class="@BuildNavigationLinkClass(NavigationSection.AccountsFinance, selectedSection) mu-finance" href="@linkGenerator.FinanceLink("AccountsFinance", accountId)">Finance</a>
                     </li>
                     <li class="das-navigation__list-item">
                         <a class="@BuildNavigationLinkClass(NavigationSection.RecruitHome, selectedSection) mu-recruit" href="@linkGenerator.RecruitLink("RecruitHome", accountId)">Recruitment</a>
